# docs/swagger.yaml
definitions:
...
  requestmodels.ChangeStatus:
    properties:
      ordered_items_id:
        type: string
      restaurantId:
        type: string
      status:
        maxLength: 20
        type: string
    required:
    - ordered_items_id
    - status
    type: object
...
  requestmodels.OrderDetails:
    properties:
      Payment_method:
        maxLength: 20
        type: string
      address_id:
        maxLength: 5
        minLength: 1
        type: string
      coupon_code:
        maxLength: 20
        type: string
      couponId:
        type: integer
      orderIdRazorPay:
        type: string
      orderStatus:
        type: string
      paymentStatus:
        type: string
      totalAmount:
        type: number
      userID:
        type: string
    required:
    - Payment_method
    - address_id
    - userID
    type: object
...
  requestmodels.RazorWebOut:
    properties:
      orderId:
        type: string
      razorpay_order_id:
        type: string
      razorpay_payment_id:
        type: string
      razorpay_signature:
        type: string
      rsecrect:
        type: string
      userId:
        type: string
    type: object
...
  requestmodels.SalesReportYYMMDD:
    properties:
      day:
        maxLength: 2
        type: string
      month:
        maxLength: 2
        type: string
      year:
        maxLength: 4
        type: string
    required:
    - year
    type: object
...
  responsemodels.Response:
    properties:
      after exicution: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: Online Veg Food Delivery.
  title: Go + Gin Veg*n.
  version: "1.0"
paths:
...
  /order:
    get:
      consumes:
      - application/json
      description: Retrieves all orders for the current user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GetAllOrders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Places a new order for the current user.
      parameters:
      - description: Order details.
        in: body
        name: orderDetails
        required: true
        schema:
          $ref: '#/definitions/requestmodels.OrderDetails'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: PlaceNewOrder
      tags:
      - Orders
  /order/invoice/{orderid}:
    get:
      consumes:
      - application/json
      description: Generates an invoice for a specific order.
      parameters:
      - description: The ID of the order to generate an invoice for.
        in: path
        name: orderid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: GenerateInvoice
      tags:
      - Orders
  /payments/verify/{orderid}:
    post:
      consumes:
      - application/json
      description: Verifies an online payment for a specific order.
      parameters:
      - description: The ID of the order to verify payment for.
        in: path
        name: orderid
        required: true
        type: string
      - description: Razorpay payment response details.
        in: body
        name: razorResp
        required: true
        schema:
          $ref: '#/definitions/requestmodels.RazorWebOut'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - UserAuthTokenAuth: []
      - UserRefTokenAuth: []
      summary: VerifyPayment
      tags:
      - Payments
...
  /restaurant/orders:
    get:
      consumes:
      - application/json
      description: Retrieves all orders for a specific restaurant.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: FetchAllOrdersForRestaurant
      tags:
      - RestaurantOrderManagement
  /restaurant/orders/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of an order.
      parameters:
      - description: Details of the status change.
        in: body
        name: changeStat
        required: true
        schema:
          $ref: '#/definitions/requestmodels.ChangeStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: ChangeStatus
      tags:
      - RestaurantOrderManagement
  /restaurant/salesreport:
    post:
      consumes:
      - application/json
      description: Generates a sales report for a specific date range or a specific
        month or a specific year.
      parameters:
      - description: YYMMDD for the report .
        in: body
        name: yymmdd
        required: true
        schema:
          $ref: '#/definitions/requestmodels.SalesReportYYMMDD'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: GetSalesreporYYMMDD
      tags:
      - RestaurantReports
  /restaurant/salesreport/{customdays}:
    get:
      consumes:
      - application/json
      description: Generates a sales report for a custom number of days.
      parameters:
      - description: Number of recent days for the report.
        in: path
        name: customdays
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: GetSalesreporCustomDays
      tags:
      - RestaurantReports
  /restaurant/salesreport/xlsx:
    get:
      consumes:
      - application/json
      description: Generates a sales report in XLSX format.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responsemodels.Response'
      security:
      - RestaurantAuthTokenAuth: []
      - RestaurantRefTokenAuth: []
      summary: GenerateSalesReportXlsx
      tags:
      - RestaurantReports
...
securityDefinitions:
  AdminRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  OtpTempTokenAuth:
    in: header
    name: Authorizations
    type: apiKey
  RestaurantAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  RestaurantRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
  UserAuthTokenAuth:
    in: header
    name: accesstoken
    type: apiKey
  UserRefTokenAuth:
    in: header
    name: refreshtoken
    type: apiKey
swagger: "2.0"